refactor: 도메인 패키지 수정 - 도메인 엔티티 별로 패키지 구성
