+++ src/main/java/nextstep/domain/QuerySetting.java
@@ -1,4 +1,4 @@
-package nextstep.domain.repository;
+package nextstep.domain;
 
 public class QuerySetting {
