+++ src/main/java/nextstep/domain/theme/Theme.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.domain.theme;
 
 public class Theme {
     private String name;
