+++ src/test/java/nextstep/service/ReservationServiceTest.java
@@ -1,8 +1,8 @@
 package nextstep.service;
 
-import nextstep.domain.Reservation;
-import nextstep.domain.Theme;
-import nextstep.domain.repository.ReservationRepository;
+import nextstep.domain.reservation.Reservation;
+import nextstep.domain.theme.Theme;
+import nextstep.domain.reservation.repository.ReservationRepository;
 import nextstep.dto.request.CreateReservationRequest;
 import nextstep.error.ApplicationException;
 import org.junit.jupiter.api.Test;
